refactor: extract useColorMode hook
